Made item a stream again. tolist kept, but discouraged.
