+++ test/test_item.py
@@ -1,9 +1,20 @@
 from stream import item, Source
+from stream.tolist import tolist
 
 def checkitem(s):
     for N in [20, 44, 51]:
         a = list(range(N))[s]
-        b = Source(range(N)) >> item[s]
+        try:
+            b = Source(range(N)) >> item[s]
+        except ValueError:
+            return
+        print(N, s)
+        assert tuple(a) == tuple(b)
+
+def checktolist(s):
+    for N in [20, 44, 51]:
+        a = list(range(N))[s]
+        b = Source(range(N)) >> tolist[s]
         print(N, s)
         assert tuple(a) == tuple(b)
 
@@ -14,3 +25,4 @@ def test_item():
         for a in stops:
             for b in stops:
                 checkitem( slice(a, b, step) )
+                checktolist( slice(a, b, step) )
